# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: |
    Tree Friend E-Commerce API.

    ## Versioning

    The API is mounted under two prefixes:
      - `/api/v1` — the canonical, versioned path. New clients should use this.
      - `/api`     — backward-compat alias for existing callers. Same handlers,
                     transparently aliased.

    When a v2 is needed, a new router will be mounted under `/api/v2` and
    `/api/v1` will be left untouched. The unversioned `/api` alias will be
    deprecated (Deprecation header + warning log) once all callers have
    migrated to `/api/v1`.
servers:
  - url: /api/v1
    description: Versioned API path (canonical)
  - url: /api
    description: Unversioned API path (backward-compat alias)
...
paths:
...
  /categories/{categoryId}/listing-attribute-options:
    get:
      operationId: listListingAttributeOptions
      tags: [sellerListings]
      summary: Get the controlled option set for a category, optionally filtered by attribute (plan doc §3a)
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
        - name: attributeName
          in: query
          required: false
          schema:
            type: string
            enum: [height, pot_size, age, root_type]
      responses:
        "200":
          description: Option values for this category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ListingAttributeOption"
  /admin/listing-attribute-options:
    post:
      operationId: createListingAttributeOption
      tags: [sellerListings]
      summary: Admin — add an option value to a category/attribute pair
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListingAttributeOptionBody"
      responses:
        "201":
          description: Created option
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingAttributeOption"
  /admin/listing-attribute-options/{id}:
    put:
      operationId: updateListingAttributeOption
      tags: [sellerListings]
      summary: Admin — update an option's value/display order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateListingAttributeOptionBody"
      responses:
        "200":
          description: Updated option
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingAttributeOption"
    delete:
      operationId: deleteListingAttributeOption
      tags: [sellerListings]
      summary: Admin — delete an option value
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /seller/orders:
    get:
      operationId: listSellerOrders
      tags: [sellerOrders]
      summary: Seller — list orders belonging to them (Manage Orders, plan doc §4)
      parameters:
        - name: orderStatus
          in: query
          required: false
          schema:
            type: string
            enum: [pending, confirmed, processing, shipped, delivered, cancelled]
      responses:
        "200":
          description: Seller's orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SellerOrder"
  /seller/orders/{id}:
    get:
      operationId: getSellerOrder
      tags: [sellerOrders]
      summary: Seller — get one of their own orders by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellerOrder"
  /seller/orders/{id}/status:
    put:
      operationId: updateSellerOrderStatus
      tags: [sellerOrders]
      summary: Seller — advance or cancel one of their own orders
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSellerOrderStatusBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellerOrder"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    OrderLineItem:
      type: object
      description: >-
        Snapshot of a purchased line item, stored as JSON on the order at
        checkout time. Either variantId+variantName are set (admin-direct
        line) or sellerListingId+sellerId are set (marketplace line) --
        never both. Every line on a given order has the same kind and, for
        marketplace orders, the same sellerId as the order itself.
      properties:
        productId:
          type: integer
        productName:
          type: string
        productImage:
          type: string
        variantId:
          type: integer
        variantName:
          type: string
        sellerListingId:
          type: integer
        sellerId:
          type: integer
        quantity:
          type: integer
        price:
          type: number
        deliveryCharge:
          type: number
      required:
        - productId
        - productName
        - productImage
        - quantity
        - price
        - deliveryCharge
...
    SellerOrder:
      type: object
      description: >-
        Seller-facing order view (Manage Orders, plan doc §4) -- a subset of
        Order's fields plus the buyer's contact and current shipment status
        folded in, omitting fields only the buyer/platform need (userId,
        senderNumber, couponCode, discountAmount, giftWrap/giftMessage).
      properties:
        id:
          type: integer
        trackingId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderLineItem"
        totalAmount:
          type: number
        paymentMethod:
          type: string
        paymentStatus:
          type: string
        orderStatus:
          type: string
        shippingAddress:
          $ref: "#/components/schemas/AddressBody"
        buyerEmail:
          type: ["string", "null"]
        cancellationReason:
          type: ["string", "null"]
        shipment:
          oneOf:
            - type: object
              properties:
                courierProvider:
                  type: string
                courierTrackingId:
                  type: ["string", "null"]
                status:
                  type: string
                lastSyncedAt:
                  type: ["string", "null"]
              required:
                - courierProvider
                - status
            - type: "null"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - trackingId
        - items
        - totalAmount
        - paymentMethod
        - paymentStatus
        - orderStatus
        - shippingAddress
        - createdAt
        - updatedAt
    UpdateSellerOrderStatusBody:
      type: object
      properties:
        orderStatus:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
        cancellationReason:
          type: string
          description: Required when orderStatus is "cancelled"
      required:
        - orderStatus
...
    AddressBody:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        street:
          type: string
        city:
          type: string
        district:
          type: string
        postalCode:
          type: ["string", "null"]
        isDefault:
          type: boolean
      required:
        - fullName
        - phone
        - street
        - city
        - district
...
    ListingAttributeOption:
      type: object
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        attributeName:
          type: string
          enum: [height, pot_size, age, root_type]
        value:
          type: string
        displayOrder:
          type: integer
      required:
        - id
        - categoryId
        - attributeName
        - value
        - displayOrder
    CreateListingAttributeOptionBody:
      type: object
      properties:
        categoryId:
          type: integer
        attributeName:
          type: string
          enum: [height, pot_size, age, root_type]
        value:
          type: string
        displayOrder:
          type: integer
      required:
        - categoryId
        - attributeName
        - value
    UpdateListingAttributeOptionBody:
      type: object
      properties:
        value:
          type: string
        displayOrder:
          type: integer
...